refactor: consolidate calendar utilities into proper folder structure
- Move calendar types (IEvent, ICalendarCell, TCalendarView, etc.) to src/types
- Move calendar helpers to src/lib/calendar-helpers.ts
- Move calendar components to src/components/calendar/
- Remove unused src/calendar folder
- Fix lint errors (convert div[role=button] to button elements)
