+++ src/app/calendar/page.tsx
@@ -1,12 +1,12 @@
 import { formatDate } from "date-fns";
 import type { Metadata } from "next";
 import { Suspense } from "react";
-import { CalendarProvider } from "@/calendar/contexts/calendar-context";
-import { monthMap } from "@/calendar/helpers";
 import { BackgroundPattern } from "@/components/background-pattern";
+import { CalendarProvider } from "@/components/calendar/contexts/calendar-context";
 import { EventsLoadingSkeleton } from "@/components/events-loading-skeleton";
 import { EventsTabs } from "@/components/events-tabs";
 import { PageHeader } from "@/components/page-header";
+import { monthMap } from "@/lib/calendar-helpers";
 import { getEvents } from "@/lib/get-events";
 import { generatePageMetadata } from "@/lib/metadata";
