+++ src/components/calendar/agenda-view/agenda-event-card.tsx
@@ -2,11 +2,11 @@
 
 import { format, parseISO } from "date-fns";
 import { Clock, MapPin } from "lucide-react";
-import { useCalendar } from "@/calendar/contexts/calendar-context";
-import type { IEvent } from "@/calendar/interfaces";
+import { useCalendar } from "@/components/calendar/contexts/calendar-context";
 import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
 import { cn, sanitizeTag } from "@/lib/utils";
+import type { IEvent } from "@/types";
 
 interface IProps {
   event: IEvent;
@@ -27,27 +27,24 @@ export function AgendaEventCard({ event, className }: IProps) {
       ? startDay
       : `${startDay}-${isSameMonth ? endDay : format(endDate, "MMM d")}`;
 
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLDivElement>) => {
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLButtonElement>) => {
     if (e.key === "Enter" || e.key === " ") {
       setSelectedEventId(event.id);
     }
   };
 
   return (
-    // biome-ignore lint/a11y/useSemanticElements: This div acts as a button for event selection.
-    <div
-      role="button"
-      tabIndex={0}
+    <button
+      type="button"
       className={cn(
         "group relative flex cursor-pointer overflow-hidden border transition-all hover:shadow-md hover:border-primary/50",
-        "hover:bg-accent/5 rounded-lg",
+        "hover:bg-accent/5 rounded-lg w-full text-left",
         className,
       )}
       onClick={() => setSelectedEventId(event.id)}
       onKeyDown={handleKeyDown}
     >
       <div className="flex w-full flex-row">
-        {/* Date Badge Section */}
         <div className="h-full flex justify-center items-center">
           <div className="m-1 flex w-24 h-24 flex-col items-center justify-center bg-background text-center rounded-lg border overflow-hidden">
             <span className="flex h-6 w-full items-center justify-center bg-primary text-center text-xs font-semibold text-primary-foreground">
@@ -59,7 +56,6 @@ export function AgendaEventCard({ event, className }: IProps) {
           </div>
         </div>
 
-        {/* Content Section */}
         <div className="flex flex-1 flex-col justify-between p-4 sm:flex-row sm:items-start sm:gap-4">
           <div className="flex flex-col gap-2">
             <div className="space-y-1">
@@ -78,12 +74,10 @@ export function AgendaEventCard({ event, className }: IProps) {
                 <MapPin className="size-3.5" />
                 <span>{event.location}</span>
               </div>
-              {/* Separator for desktop */}
               <Separator
                 orientation="vertical"
                 className="hidden h-3 sm:block"
               />
-              {/* Tags */}
               {event.tags && event.tags.length > 0 && (
                 <div className="flex flex-wrap gap-1.5">
                   {event.tags.slice(0, 3).map((tag) => (
@@ -100,7 +94,6 @@ export function AgendaEventCard({ event, className }: IProps) {
             </div>
           </div>
 
-          {/* Right Side: Time */}
           <div className="mt-4 flex items-center justify-end gap-1.5 text-sm font-medium text-primary sm:mt-0 sm:flex-col sm:items-end sm:justify-start sm:gap-0">
             <Clock className="size-4 sm:hidden" />
             <span className="whitespace-nowrap">
@@ -109,6 +102,6 @@ export function AgendaEventCard({ event, className }: IProps) {
           </div>
         </div>
       </div>
-    </div>
+    </button>
   );
 }
