+++ src/components/timeline.tsx
@@ -3,11 +3,11 @@
 import { format } from "date-fns";
 import { motion } from "framer-motion";
 import { useState } from "react";
-import { EventDetailsDialog } from "@/calendar/components/dialogs/event-details-dialog";
-import type { IEvent } from "@/calendar/interfaces";
+import { EventDetailsDialog } from "@/components/calendar/dialogs/event-details-dialog";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent } from "@/components/ui/card";
 import { iconsMap } from "@/constants";
+import type { IEvent } from "@/types";
 
 interface TimelineProps {
   timelineEvents: IEvent[];
