+++ src/lib/get-events.ts
@@ -5,10 +5,9 @@ import path from "node:path";
 import { endOfMonth } from "date-fns";
 import { JWT } from "google-auth-library";
 import { GoogleSpreadsheet } from "google-spreadsheet";
-import type { IEvent } from "@/calendar/interfaces";
 import type { iconsMap } from "@/constants";
 import type { EventSubmissionFormValues } from "@/lib/forms-config";
-import type { TEventColor } from "@/types";
+import type { IEvent, TEventColor } from "@/types";
 
 const serviceAccountAuth = new JWT({
   email: process.env.GOOGLE_SERVICE_ACCOUNT_EMAIL,
