+++ src/types/index.ts
@@ -9,6 +9,43 @@ export type TEventColor =
   | "orange"
   | "gray";
 
+export interface IUser {
+  id: string;
+  name: string;
+  picturePath: string | null;
+}
+
+export interface IEvent {
+  id: string;
+  title: string;
+  startDateTime: string;
+  endDateTime: string;
+  time: string;
+  location: string;
+  description: string;
+  registrationLink: string;
+  joinLink: string;
+  icon: keyof typeof iconsMap;
+  highlight: boolean;
+  youtubeLink?: string;
+  color?: TEventColor;
+  organizationName?: string;
+  posterUrl?: string;
+  tags?: string[];
+}
+
+export interface ICalendarCell {
+  day: number;
+  currentMonth: boolean;
+  date: Date;
+}
+
+export type TCalendarView = "day" | "week" | "month" | "year" | "agenda";
+
+export type TBadgeVariant = "dot" | "colored" | "mixed";
+export type TWorkingHours = { [key: number]: { from: number; to: number } };
+export type TVisibleHours = { from: number; to: number };
+
 export interface CommunityRepresentative {
   name: string;
   role?: string;
